backend/requirements.txt
fastapi
uvicorn
vaderSentiment
nltk
requests
beautifulsoup4
google-genai
python-dotenv